Sleep more often during FAQs import.
I kept getting a 503 error whenever I ran the import in staging. I
figured that the sleep added by Tim must have been to avoid hitting the
same problem in dev. Not sure why full imports work in dev without this
change but not in staging, but after testing this in staging the import
worked.