Make Fbopen full data importer memory friendly
Parse the full dump as we iterate thru it instead of loading all entries
in one go.
It should use roughly the same amount of memory throughout the import.
And it should be a lot less than before.

This imported the 1GB dump xml in 1 hour on a 1GB vps while previously it
would get oom killed in the first 10 minutes :-)